requirements.txt
# WhatsApp Bot Dependencies
# Core web scraping with HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Email sending via Resend
resend>=0.6.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP session management and parsing
urllib3>=2.0.0
html5lib>=1.1

# Optional: For enhanced logging and scheduling
schedule>=1.2.0
colorama>=0.4.6

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
